Struct sc_chain_spec::GenericChainSpec

pub struct GenericChainSpec<G, E = NoExtension> { /* private fields */ }

A configuration of a chain. Can be used to build a genesis block.

Implementations

impl<G, E> ChainSpec<G, E>

pub fn boot_nodes(&self) -> &[MultiaddrWithPeerId]

A list of bootnode addresses.

pub fn name(&self) -> &str

Spec name.

pub fn id(&self) -> &str

Spec id.

pub fn telemetry_endpoints(&self) -> &Option<TelemetryEndpoints>

Telemetry endpoints (if any)

pub fn protocol_id(&self) -> Option<&str>

Network protocol id.

pub fn fork_id(&self) -> Option<&str>

Optional network fork identifier.

pub fn properties(&self) -> Properties

Additional loosly-typed properties of the chain.

Returns an empty JSON object if ‘properties’ not defined in config

pub fn add_boot_node(&mut self, addr: MultiaddrWithPeerId)

Add a bootnode to the list.

pub fn extensions(&self) -> &E

Returns a reference to the defined chain spec extensions.

pub fn extensions_mut(&mut self) -> &mut E

Returns a mutable reference to the defined chain spec extensions.

pub fn from_genesis<F: Fn() -> G + 'static + Send + Sync>( name: &str, id: &str, chain_type: ChainType, constructor: F, boot_nodes: Vec<MultiaddrWithPeerId>, telemetry_endpoints: Option<TelemetryEndpoints>, protocol_id: Option<&str>, fork_id: Option<&str>, properties: Option<Properties>, extensions: E ) -> Self

Create hardcoded spec.

impl<G, E: DeserializeOwned> ChainSpec<G, E>

pub fn from_json_bytes( json: impl Into<Cow<'static, [u8]>> ) -> Result<Self, String>

Parse json content into a ChainSpec

pub fn from_json_file(path: PathBuf) -> Result<Self, String>

Parse json file into a ChainSpec

impl<G: RuntimeGenesis, E: Serialize + Clone + 'static> ChainSpec<G, E>

pub fn as_json(&self, raw: bool) -> Result<String, String>

Dump to json string.