sc_chain_spec

Trait BuildGenesisBlock

pub trait BuildGenesisBlock<Block: BlockT> {
    type BlockImportOperation;

    // Required method
    fn build_genesis_block(self) -> Result<(Block, Self::BlockImportOperation)>;
}

Trait for building the genesis block.

Required Associated Types

type BlockImportOperation

The import operation used to import the genesis block into the backend.

Required Methods

fn build_genesis_block(self) -> Result<(Block, Self::BlockImportOperation)>

Returns the built genesis block along with the block import operation after setting the genesis storage.