Trait sc_chain_spec::ChainSpec

pub trait ChainSpec: BuildStorage + Send + Sync {

    // Required methods
    fn name(&self) -> &str;
    fn id(&self) -> &str;
    fn chain_type(&self) -> ChainType;
    fn boot_nodes(&self) -> &[MultiaddrWithPeerId];
    fn telemetry_endpoints(&self) -> &Option<TelemetryEndpoints>;
    fn protocol_id(&self) -> Option<&str>;
    fn fork_id(&self) -> Option<&str>;
    fn properties(&self) -> Properties;
    fn extensions(&self) -> &dyn GetExtension;
    fn extensions_mut(&mut self) -> &mut dyn GetExtension;
    fn add_boot_node(&mut self, addr: MultiaddrWithPeerId);
    fn as_json(&self, raw: bool) -> Result<String, String>;
    fn as_storage_builder(&self) -> &dyn BuildStorage;
    fn cloned_box(&self) -> Box<dyn ChainSpec>;
    fn set_storage(&mut self, storage: Storage);
    fn code_substitutes(&self) -> BTreeMap<String, Vec<u8>>;

}

Common interface of a chain specification.

Required Methods

fn name(&self) -> &str

Spec name.

fn id(&self) -> &str

Spec id.

fn chain_type(&self) -> ChainType

Type of the chain.

fn boot_nodes(&self) -> &[MultiaddrWithPeerId]

A list of bootnode addresses.

fn telemetry_endpoints(&self) -> &Option<TelemetryEndpoints>

Telemetry endpoints (if any)

fn protocol_id(&self) -> Option<&str>

Network protocol id.

fn fork_id(&self) -> Option<&str>

Optional network fork identifier. None by default.

fn properties(&self) -> Properties

Additional loosely-typed properties of the chain.

Returns an empty JSON object if ‘properties’ not defined in config

fn extensions(&self) -> &dyn GetExtension

Returns a reference to the defined chain spec extensions.

fn extensions_mut(&mut self) -> &mut dyn GetExtension

Returns a mutable reference to the defined chain spec extensions.

fn add_boot_node(&mut self, addr: MultiaddrWithPeerId)

Add a bootnode to the list.

fn as_json(&self, raw: bool) -> Result<String, String>

Return spec as JSON.

fn as_storage_builder(&self) -> &dyn BuildStorage

Return StorageBuilder for this spec.

fn cloned_box(&self) -> Box<dyn ChainSpec>

Returns a cloned Box<dyn ChainSpec>.

fn set_storage(&mut self, storage: Storage)

Set the storage that should be used by this chain spec.

This will be used as storage at genesis.

fn code_substitutes(&self) -> BTreeMap<String, Vec<u8>>

Returns code substitutes that should be used for the on chain wasm.