Trait sc_chain_spec::Fork

pub trait Fork: Serialize + DeserializeOwned + Clone + Sized {
    type Base: Group<Fork = Self>;

    // Required methods
    fn combine_with(&mut self, other: Self);
    fn to_base(self) -> Option<Self::Base>;
}

A ChainSpec extension fork definition.

Basically should look the same as Group, but all parameters are optional. This allows changing only one parameter as part of the fork. The forks can be combined (summed up) to specify a complete set of parameters

Required Associated Types

type Base: Group<Fork = Self>

A base Group type.

Required Methods

fn combine_with(&mut self, other: Self)

Combine with another struct.

All parameters set in other should override the ones in the current struct.

fn to_base(self) -> Option<Self::Base>

Attempt to convert to the base type if all parameters are set.

Object Safety

This trait is not object safe.